+++ alasql.js
@@ -3997,6 +3997,32 @@ alasql.matrix = alasql.queryArrayOfArrays;
 */
 
 
+//
+// Promises for AlaSQL
+//
+
+if(typeof exports == 'object') {
+	var Promise = require('es6-promise').Promise;
+} 
+
+//
+// Only for browsers with Promise support
+//
+
+if(typeof Promise == 'function') {
+	alasql.async = function(sql, params) {
+	    return new Promise(function(resolve, reject){
+	        alasql(sql, params, function(data,err) {
+	             if(err) {
+	                 reject(err);
+	             } else {
+	                 resolve(data);
+	             }
+	        });
+	    });
+	};	
+}
+
 /*
 //
 // Database class for Alasql.js