+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ### 0.1.6 "Palermo" (13.05.2015 - ...05.2015)
 * SET NOCOUNT OFF (for CREATE and INSERT)
 * ROWNUM() and ROW_NUMBER() functions
-*
+* Promised version of alasql() - alasql.async() (based on es6-promises)
+* SELECT * FROM Json
 
 ### 0.1.5 "San Marino" (12.05.2015 - 12.05.2015)
 * Added Meteor package (agershun:alasql) - still does not work - skeleton